# expense-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Expense Management API",
        "description": "Multi-Tenant SaaS Expense Management API — secure, role-based, company-scoped.",
        "contact": {
            "email": "gplanet.tech@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Local / Development"
        },
        {
            "url": "https://api.yourdomain.com/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/expenses": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "List expenses (paginated, searchable, company-scoped)",
                "operationId": "36e62ecfbcfde37686ea63fe89d0c0c0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search title or category",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Exact category filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "created_at >=",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "created_at <=",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "amount",
                                "title",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated expense list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Create an expense (company + user auto-derived from token)",
                "operationId": "927e29ac31ae983af0a55ce5d487d554",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "amount",
                                    "category"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Team lunch"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 85.5
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Food"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created — audit log written"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/expenses/{id}": {
            "get": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Get a single expense",
                "operationId": "b8d5c673ecdbb6a51bea7ec7cc5d0e30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense detail"
                    },
                    "403": {
                        "description": "Unauthorized (policy)"
                    },
                    "404": {
                        "description": "Not found or cross-company"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Update an expense (Manager|Admin only — audit-logged)",
                "operationId": "36a4200b139cad8582ff017bbdb5750e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "category": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated"
                    },
                    "403": {
                        "description": "Insufficient permissions"
                    },
                    "404": {
                        "description": "Not found or cross-company"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Expenses"
                ],
                "summary": "Delete an expense (Admin only — audit-logged)",
                "operationId": "b261a1443844b958952ae83b03fbbfa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense deleted"
                    },
                    "403": {
                        "description": "Insufficient permissions — Admin required"
                    },
                    "404": {
                        "description": "Not found or cross-company"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum token — obtain via /auth/login or /auth/register",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}